feat: Add Fiber framework for web server and routes
The code changes include adding the Fiber framework for creating a web server and defining routes. This will allow the application to handle HTTP requests and responses. The recent user commits indicate the addition of dependencies for Fiber and PostgreSQL, as well as the creation of a Docker compose file for containerization. The recent repository commits also show the addition of Fiber dependencies and updates to the README file.